注册提交

+++ com.jebao.root/3.1-com.jebao.jebaodb.dao/src/main/java/com/jebao/jebaodb/dao/mapper/user/TbLoginInfoMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import java.util.List;
 
 public interface TbLoginInfoMapper {
-    int insert(TbLoginInfo record);
+    long insert(TbLoginInfo record);
 
     int insertSelective(TbLoginInfo record);
